+++ compiler/syntax/src/res_parens.ml
@@ -271,9 +271,9 @@ let field_expr expr =
     | {
      pexp_desc =
        ( Pexp_assert _ | Pexp_extension _ (* %extension.x vs (%extension).x *)
-       | Pexp_object_literal _ (* ({"a": 1})["a"] *) | Pexp_fun _
-       | Pexp_constraint _ | Pexp_coerce _ | Pexp_setfield _ | Pexp_match _
-       | Pexp_try _ | Pexp_while _ | Pexp_for _ | Pexp_for_of _
+       | Pexp_regexp _ | Pexp_object_literal _ (* ({"a": 1})["a"] *)
+       | Pexp_fun _ | Pexp_constraint _ | Pexp_coerce _ | Pexp_setfield _
+       | Pexp_match _ | Pexp_try _ | Pexp_while _ | Pexp_for _ | Pexp_for_of _
        | Pexp_for_await_of _ | Pexp_ifthenelse _ );
     } ->
       Parenthesized
@@ -365,11 +365,11 @@ let jsx_child_expr expr =
       | _ when Parsetree_viewer.expr_is_await expr -> Parenthesized
       | {
        Parsetree.pexp_desc =
-         ( Pexp_ident _ | Pexp_constant _ | Pexp_field _ | Pexp_construct _
-         | Pexp_variant _ | Pexp_array _ | Pexp_pack _ | Pexp_record _
-         | Pexp_object_literal _ | Pexp_extension _ | Pexp_letmodule _
-         | Pexp_letexception _ | Pexp_open _ | Pexp_sequence _ | Pexp_let _
-         | Pexp_jsx_element _ );
+         ( Pexp_ident _ | Pexp_constant _ | Pexp_regexp _ | Pexp_field _
+         | Pexp_construct _ | Pexp_variant _ | Pexp_array _ | Pexp_pack _
+         | Pexp_record _ | Pexp_object_literal _ | Pexp_extension _
+         | Pexp_letmodule _ | Pexp_letexception _ | Pexp_open _
+         | Pexp_sequence _ | Pexp_let _ | Pexp_jsx_element _ );
        pexp_attributes = [];
       } ->
         Nothing