+++ compiler/syntax/src/res_printer.ml
@@ -3471,20 +3471,9 @@ and print_expression ~state (e : Parsetree.expression) cmt_tbl =
              Doc.soft_line;
              Doc.rbrace;
            ])
-    | Pexp_extension extension -> (
-      match extension with
-      | ( {txt = "re"},
-          PStr
-            [
-              {
-                pstr_desc =
-                  Pstr_eval
-                    ({pexp_desc = Pexp_constant (Pconst_raw_source expr)}, []);
-              };
-            ] ) ->
-        Doc.text expr
-      | extension ->
-        print_extension ~state ~at_module_lvl:false extension cmt_tbl)
+    | Pexp_extension extension ->
+      print_extension ~state ~at_module_lvl:false extension cmt_tbl
+    | Pexp_regexp {pattern; flags} -> Doc.text ("/" ^ pattern ^ "/" ^ flags)
     | Pexp_template {source_segments; values} ->
       print_template_literal ~state ~source_segments ~values cmt_tbl
     | Pexp_tagged_template {tag; raw_sources; values} ->